Rename providers to handlers
