+++ src/modules/Elsa.Http/DownloadableContentHandlers/BinaryDownloadableContentHandler.cs
@@ -2,12 +2,12 @@ using Elsa.Http.Abstractions;
 using Elsa.Http.Contexts;
 using Elsa.Http.Models;
 
-namespace Elsa.Http.DownloadableProviders;
+namespace Elsa.Http.DownloadableContentHandlers;
 
 /// <summary>
 /// Handles content that represents a downloadable binary file.
 /// </summary>
-public class BinaryDownloadableProvider : DownloadableProviderBase
+public class BinaryDownloadableContentHandler : DownloadableContentHandlerBase
 {
     /// <inheritdoc />
     public override bool GetSupportsContent(object content) => content is byte[];
