+++ src/modules/Elsa.Http/DownloadableContentHandlers/StreamDownloadableContentHandler.cs
@@ -2,12 +2,12 @@ using Elsa.Http.Abstractions;
 using Elsa.Http.Contexts;
 using Elsa.Http.Models;
 
-namespace Elsa.Http.DownloadableProviders;
+namespace Elsa.Http.DownloadableContentHandlers;
 
 /// <summary>
 /// Handles content that represents a downloadable stream.
 /// </summary>
-public class StreamDownloadableProvider : DownloadableProviderBase
+public class StreamDownloadableContentHandler : DownloadableContentHandlerBase
 {
     /// <inheritdoc />
     public override bool GetSupportsContent(object content) => content is Stream;
